// api/models/bookChapter.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */

/**
 * A book chapter. Includes the title and timestamps.
 * @nullable
 */
export type BookChapter = {
  /** The ID of the book chapter. */
  id?: number;
  /** When in the book (in seconds) the chapter starts. */
  start?: number;
  /** When in the book (in seconds) the chapter ends. */
  end?: number;
  /** The title of the chapter. */
  title?: string;
} | null;